pyproject.toml
[tool.poetry]
name = "vai-technologies-test-task"
version = "0.1.0"
description = ""
authors = ["nabas <a6ashin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.7.2"
build = "1.1.1"
cachecontrol = "0.14.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
colorama = "0.4.6"
crashtest = "0.4.1"
distlib = "0.3.8"
django = "5.0.3"
django-filter = "24.1"
djangorestframework = "3.14.0"
drf-yasg = "1.21.7"
dulwich = "0.21.7"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
idna = "3.6"
importlib-metadata = "7.0.2"
inflection = "0.5.1"
installer = "0.7.0"
jaraco-classes = "3.3.1"
keyring = "24.3.1"
more-itertools = "10.2.0"
msgpack = "1.0.8"
packaging = "24.0"
pexpect = "4.9.0"
pkginfo = "1.10.0"
platformdirs = "4.2.0"
poetry = "1.8.2"
poetry-core = "1.9.0"
poetry-plugin-export = "1.6.0"
ptyprocess = "0.7.0"
pyproject-hooks = "1.0.0"
pytz = "2024.1"
pywin32-ctypes = "0.2.2"
pyyaml = "6.0.1"
rapidfuzz = "3.6.2"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
sqlparse = "0.4.4"
tomlkit = "0.12.4"
trove-classifiers = "2024.3.3"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
virtualenv = "20.25.1"
zipp = "3.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
